+++ assets/scripts/segments/drag_and_drop.js
@@ -8,8 +8,7 @@ import { infoBubble } from '../info_bubble/info_bubble'
 import { hideAllMenus } from '../menus/menu_controller'
 import { app } from '../preinit/app_settings'
 import { system } from '../preinit/system_capabilities'
-import { getStreet } from '../streets/data_model'
-import { setIgnoreStreetChanges } from '../streets/undo_stack'
+import { getStreet, setIgnoreStreetChanges } from '../streets/data_model'
 import { getElAbsolutePos } from '../util/helpers'
 import { generateRandSeed } from '../util/random'
 import { BUILDING_SPACE } from './buildings'
