+++ assets/scripts/segments/resizing.js
@@ -1,8 +1,7 @@
 import { trackEvent } from '../app/event_tracking'
 import { INFO_BUBBLE_TYPE_SEGMENT, infoBubble } from '../info_bubble/info_bubble'
 import { system } from '../preinit/system_capabilities'
-import { getStreet } from '../streets/data_model'
-import { setIgnoreStreetChanges } from '../streets/undo_stack'
+import { getStreet, setIgnoreStreetChanges } from '../streets/data_model'
 import {
   SEGMENT_WARNING_OUTSIDE,
   SEGMENT_WARNING_WIDTH_TOO_SMALL,
